Add files via upload
A stroke is an interruption of the blood supply to any part of the brain. 
It is also referred to as a brain attack. Stroke is the third leading cause of death and the leading cause of adult disability.
According to the World Health Organization (WHO) stroke is the 2nd leading cause of death globally, responsible for approximately 11% of total deaths.Early identification of stroke can help doctors to give necessary medication to the patient
The machine learning model (XGBoost) was trained and validated on 5100+  patient health records
Achieved True Positive rate of 0.97 and F1 score of 0.97